More renaming and rearranging
