+++ src/lib.rs
@@ -127,7 +127,7 @@ mod new;
 pub mod new2;
 pub mod pool_item;
 mod pool_thread;
-mod pool_thread_2;
+mod pool_thread_old;
 mod receive;
 pub mod request_response_pair;
 mod send;
