+++ src/pool_item/pool_item_api.rs
@@ -1,0 +1,6 @@
+pub trait PoolItemApi {
+    fn is_request(&self) -> bool;
+    fn is_response(&self) -> bool {
+        !self.is_request()
+    }
+}
