+++ src/samples/mod.rs
@@ -2,8 +2,7 @@ mod randoms;
 mod randoms_batch;
 
 // re-export
-pub use randoms::randoms_request::*;
-pub use randoms::randoms_response::*;
+pub use randoms::randoms_api::*;
 pub use randoms::Randoms;
 pub use randoms_batch::randoms_batch_request::*;
 pub use randoms_batch::randoms_batch_response::*;
