+++ src/samples/randoms/mod.rs
@@ -1,8 +1,5 @@
-pub mod message_processor;
 pub mod pool_item;
 pub mod randoms_api;
-pub mod randoms_request;
-pub mod randoms_response;
 
 use rand::{RngCore, SeedableRng};
 use rand_xoshiro::Xoshiro256Plus;
