+++ src/samples/randoms/randoms_api/randoms_init_request.rs
